refactor: rename config to app_config
