+++ src/config/mod.rs
@@ -1,1 +1,1 @@
-pub mod config;
+pub mod app_config;
